remove: delete model-budget.ts and all associated tests
- shared/ai/model-budget.ts
- tests/unit/model-loader.test.ts
- tests/integration/memory-budget.test.ts
- tests/integration/model-loading.test.ts

TS compiles clean; 188 tests pass, 0 fail.
