+++ bunfig.toml
@@ -1,8 +1,6 @@
 [test]
-# Preload setup file to initialize mocks and environment before running Bun tests
+root = "./tests"
 preload = ["./tests/setup.ts"]
-
-# Enable coverage reporting and set defaults
 coverage = true
 coverageDir = "./coverage"
 timeout = 5000
