+++ expo-router/entry
@@ -1,0 +1,1 @@
+console.log("Hello via Bun!");
